+++ product-service/src/main/java/com/brunomilitzer/productservice/repository/ProductRepository.java
@@ -6,10 +6,8 @@ import org.springframework.data.mongodb.repository.ReactiveMongoRepository;
 import org.springframework.stereotype.Repository;
 import reactor.core.publisher.Flux;
 
-import java.math.BigDecimal;
-
 @Repository
 public interface ProductRepository extends ReactiveMongoRepository<Product, String> {
 
-    Flux<Product> findByPriceBetween(Range<BigDecimal> range);
+    Flux<Product> findByPriceBetween(Range<Integer> range);
 }
